fix: resolve PR issues
