+++ apps/frontend/src/app/3_organisms/ConversionsHistoryFrame/ConvertHistory.tsx
@@ -2,8 +2,8 @@ import React, { FC, useMemo, useState } from 'react';
 
 import { Select } from '@sovryn/ui';
 
+import { CONVERT_HISTORY_OPTIONS } from './ConvertHistory.constants';
 import { ConvertHistoryType } from './ConvertHistory.types';
-import { convertHistoryOptions } from './ConvertHistory.utils';
 import { AmmConversionsHistoryFrame } from './components/AmmConversionsHistoryFrame/AmmConversionsHistoryFrame';
 import { MyntConversionsHistoryFrame } from './components/MyntConversionsHistoryFrame/MyntConversionsHistoryFrame';
 import { ZeroConversionsHistoryFrame } from './components/ZeroConversionsHistoryFrame/ZeroConversionsHistoryFrame';
@@ -13,13 +13,13 @@ export const ConvertHistory: FC = () => {
     ConvertHistoryType.AMM,
   );
 
-  const selectComponent = useMemo(
+  const SelectComponent = useMemo(
     () => (
       <Select
         dataAttribute={`convert-history-${selectedHistoryType}`}
         value={selectedHistoryType}
-        onChange={(value: ConvertHistoryType) => setSelectedHistoryType(value)}
-        options={convertHistoryOptions}
+        onChange={setSelectedHistoryType}
+        options={CONVERT_HISTORY_OPTIONS}
         className="min-w-36"
       />
     ),
@@ -31,25 +31,25 @@ export const ConvertHistory: FC = () => {
       case ConvertHistoryType.AMM:
         return (
           <AmmConversionsHistoryFrame>
-            {selectComponent}
+            {SelectComponent}
           </AmmConversionsHistoryFrame>
         );
       case ConvertHistoryType.MYNT:
         return (
           <MyntConversionsHistoryFrame>
-            {selectComponent}
+            {SelectComponent}
           </MyntConversionsHistoryFrame>
         );
       case ConvertHistoryType.ZERO:
         return (
           <ZeroConversionsHistoryFrame>
-            {selectComponent}
+            {SelectComponent}
           </ZeroConversionsHistoryFrame>
         );
       default:
         return null;
     }
-  }, [selectedHistoryType, selectComponent]);
+  }, [selectedHistoryType, SelectComponent]);
 
   return <>{renderHistoryFrame}</>;
 };
