+++ apps/frontend/src/app/3_organisms/ConversionsHistoryFrame/components/MyntConversionsHistoryFrame/MyntConversionsHistoryFrame.tsx
@@ -15,11 +15,8 @@ import {
   OrderOptions,
   OrderDirection,
   NotificationType,
-  ErrorBadge,
-  ErrorLevel,
 } from '@sovryn/ui';
 
-import { ExportCSV } from '../../../../2_molecules/ExportCSV/ExportCSV';
 import { masset } from '../../../../5_pages/ConvertPage/ConvertPage.types';
 import {
   DEFAULT_HISTORY_FRAME_PAGE_SIZE,
@@ -28,7 +25,6 @@ import {
 import { useNotificationContext } from '../../../../../contexts/NotificationContext';
 import { useAccount } from '../../../../../hooks/useAccount';
 import { useBlockNumber } from '../../../../../hooks/useBlockNumber';
-import { useMaintenance } from '../../../../../hooks/useMaintenance';
 import { translations } from '../../../../../locales/i18n';
 import {
   Conversion,
@@ -36,10 +32,9 @@ import {
   useGetUserConversionsLazyQuery,
 } from '../../../../../utils/graphql/mynt/generated';
 import { dateFormat } from '../../../../../utils/helpers';
-import {
-  columnsConfig,
-  generateRowTitle,
-} from './MyntConversionsHistoryFrame.utils';
+import { BaseConversionsHistoryFrame } from '../BaseConversionsHistoryFrame/BaseConversionsHistoryFrame';
+import { COLUMNS_CONFIG } from './MyntConversionsHistoryFrame.constants';
+import { generateRowTitle } from './MyntConversionsHistoryFrame.utils';
 import { useGetMyntConversionsHistory } from './hooks/useGetMyntConversionsHistory';
 
 const pageSize = DEFAULT_HISTORY_FRAME_PAGE_SIZE;
@@ -127,48 +122,35 @@ export const MyntConversionsHistoryFrame: React.FC<PropsWithChildren> = ({
     }));
   }, [account, addNotification, getConversions]);
 
-  const { checkMaintenance, States } = useMaintenance();
-  const exportLocked = checkMaintenance(States.ZERO_EXPORT_CSV);
-
   return (
-    <>
-      <div className="flex-row items-center gap-4 mb-7 flex justify-center lg:justify-start">
-        {children}
-        <div className="flex-row items-center ml-2 gap-4 hidden lg:inline-flex">
-          <ExportCSV
-            getData={exportData}
-            filename="conversions"
-            disabled={!conversions || exportLocked}
-          />
-          {exportLocked && (
-            <ErrorBadge
-              level={ErrorLevel.Warning}
-              message={t(translations.maintenanceMode.featureDisabled)}
-            />
-          )}
-        </div>
-      </div>
-      <div className="bg-gray-80 py-4 px-4 rounded">
+    <BaseConversionsHistoryFrame
+      exportData={exportData}
+      name="Mynt-conversions"
+      table={
         <Table
           setOrderOptions={setOrderOptions}
           orderOptions={orderOptions}
-          columns={columnsConfig}
+          columns={COLUMNS_CONFIG}
           rows={conversions}
           rowTitle={generateRowTitle}
           isLoading={loading}
           className="bg-gray-80 text-gray-10 lg:px-6 lg:py-4"
           noData={t(translations.common.tables.noData)}
-          dataAttribute="mynt-conversions-history-table"
+          dataAttribute="amm-conversions-history-table"
         />
+      }
+      pagination={
         <Pagination
           page={page}
           className="lg:pb-6 mt-3 lg:mt-6 justify-center lg:justify-start"
           onChange={onPageChange}
           itemsPerPage={pageSize}
           isNextButtonDisabled={isNextButtonDisabled}
-          dataAttribute="mynt-conversions-history-pagination"
+          dataAttribute="amm-conversions-history-pagination"
         />
-      </div>
-    </>
+      }
+    >
+      {children}
+    </BaseConversionsHistoryFrame>
   );
 };
